refactor: Move ggshield.sca to ggshield.verticals.sca
This is one of the step needed to regroup all verticals into one
package and simplify the identification of the verticals layer as
defined in #521.

refs: #521
