+++ ggshield/cmd/sca/scan/diff.py
@@ -17,7 +17,9 @@ from ggshield.cmd.sca.scan.scan_common_options import (
     add_sca_scan_common_options,
     update_context,
 )
-from ggshield.sca.collection.collection import SCAScanDiffVulnerabilityCollection
+from ggshield.verticals.sca.collection.collection import (
+    SCAScanDiffVulnerabilityCollection,
+)
 
 
 @click.command()
