+++ ggshield/cmd/sca/scan/precommit.py
@@ -14,7 +14,7 @@ from ggshield.cmd.sca.scan.scan_common_options import (
     add_sca_scan_common_options,
     update_context,
 )
-from ggshield.sca.collection.collection import (
+from ggshield.verticals.sca.collection.collection import (
     SCAScanAllVulnerabilityCollection,
     SCAScanDiffVulnerabilityCollection,
 )
