+++ ggshield/verticals/sca/output/__init__.py
