+++ ggshield/verticals/sca/sca_scan_models.py
