+++ ggshield/verticals/sca/vuln_identifier.py
